jenkins button

+++ Jenkinsfile
@@ -74,13 +74,13 @@ podTemplate(label: label, containers: [
 
                 stage('Apply Terraform') {
                     container('terraform'){
-                         sh 'terraform destroy -auto-approve -input=false'
-                         //if (${params.apply} == 'terraform apply') {
-                         //   sh 'terraform apply -auto-approve -input=false myplan'
-                         //}
-                         //else {
-                         //   sh 'terraform destroy -auto-approve -input=false'
-                         //}
+                         if (params.apply == 'terraform apply') {
+                            echo "hello"
+                            sh 'terraform apply -auto-approve -input=false myplan'
+                         }
+                         else {
+                            sh 'terraform destroy -auto-approve -input=false'
+                         }
                     }
                 }
